Establishes base and nested service implementation
This might not be necessary, for the initial discovery.  But decided to follow precedence in Vet360, EVSS, etc.
