+++ lib/ihub/service.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'common/client/base'
+
+module IHub
+  class Service < Common::Client::Base
+    STATSD_KEY_PREFIX = 'api.ihub'
+
+    def initialize(user)
+      @user = user
+    end
+  end
+end
